requirements.txt
beautifulsoup4
requests
lxml
psycopg2
sqlalchemy
python-dateutil
